Change link to point to English version of kystverket.no
This commit updates the link in the README from kystverket.no to kystverket.no/en. With this change, the default language on the target website is now English, replacing the previous default language of Norwegian. This modification ensures that users are directed to the English version of the website for better accessibility and understanding.